Don't install Java if already installed

+++ provision.sh
@@ -1,2 +1,5 @@
-aptitude update
-aptitude install -y default-jre
+#!/bin/bash
+if [ ! -e /usr/bin/java ]; then
+	aptitude update
+	aptitude install -y default-jre
+fi
